Change auto pass into toggle
Change auto pass function so that it is not a pass action. Now players
can enable auto pass and then take any action, pass or otherwise. Then
it will auto pass following turns unless cancelled. Also fixed the
button to be translatable and added locking and log messages to calls.