Fix hand cards
Previous commit introduced error in hand counts when Show Hands option
was not enabled by always returning zero. Fix with correct count
regardless of options. Also initialize empty array for card backs so
that it is always present and will not throw errors accessing a null
object on client side.